requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
sentence-transformers>=2.2.0
numpy>=1.24.0
faiss-cpu>=1.7.4
googlesearch-python>=1.2.3
ddgs>=0.0.1
colorama>=0.4.6
tqdm>=4.66.0

# Local LLM support (optional but recommended for better answers)
transformers>=4.36.0
torch>=2.1.0
accelerate>=0.25.0
